requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
python-dotenv
gunicorn
